+++ paymentsheet/src/main/java/com/stripe/android/paymentelement/embedded/content/EmbeddedStateHelper.kt
@@ -71,6 +71,14 @@ internal class DefaultEmbeddedStateHelper @Inject constructor(
                     "when Google Pay or a customer configuration is provided. " +
                     "Use RowSelectionBehavior.Default or disable Google Pay and saved payment methods."
             )
+
+        }
+        if (configuration.embeddedViewDisplaysMandateText && isRowSelectionImmediateAction) {
+            throw IllegalArgumentException(
+                "Your integration must set `embeddedViewDisplaysMandateText` to false and display the mandate " +
+                    "(`embeddedPaymentElement.paymentOption.mandateText`) to customer near your buy button " +
+                    "and/or before confirmation when `RowSelectionBehavior = ImmediateAction`"
+            )
         }
     }